trainer/requirements.txt
unsloth @ git+https://github.com/unslothai/unsloth.git
unsloth_zoo @ git+https://github.com/unslothai/unsloth_zoo.git
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.36.0
trl>=0.7.0
datasets>=2.16.0
huggingface_hub>=0.19.0
python-dotenv>=1.0.0
